g++ changes
Fixed err/war when compiled with g++
Replace GMRFLib_strdup with strdup

Need to allow NULL ptr to strdup

Change from macro to function

+++ inlaprog/src/inla-special-functions.c
@@ -155,5 +155,5 @@ double inla_lgamma_fast(double x)
 double inla_ipow(double x, int k)
 {
 	// x^k
-	return gsl_sf_pow_int((const double) x, (const int) k);
+	return gsl_sf_pow_int(x, k);
 }
